backend/requirements.txt
fastapi
sqlalchemy
openai
pydantic
python-dotenv
httpx
psycopg2-binary